+++ packages/shared/src/constants/index.ts
@@ -41,4 +41,4 @@ export const ZOOM_STEP = 0.1;
 
 export const DEFAULT_ADMIN_EMAIL = 'admin@tactihub.local';
 
-export const APP_VERSION = '1.6.0';
+export const APP_VERSION = '1.7.0';